+++ src/cli/cli.ts
@@ -11,13 +11,13 @@ import { anthropic } from "@ai-sdk/anthropic";
 import { intro, log, outro } from "@clack/prompts";
 import { createProviderRegistry } from "ai";
 
-import { MarkdownAI } from "./chat.js";
-import { tryCatch } from "./utils.js";
-import { createReadFilesTool } from "./tools/read-files.js";
-import { createListFilesTool } from "./tools/list-files.js";
-import { createWriteFilesTool } from "./tools/write-files.js";
-import { createGrepSearchTool } from "./tools/grep-search.js";
-import { createExecCommandTool } from "./tools/exec-command.js";
+import { MarkdownAI } from "../chat/chat.js";
+import { tryCatch } from "../utils/index.js";
+import { createReadFilesTool } from "../tools/read-files.js";
+import { createListFilesTool } from "../tools/list-files.js";
+import { createWriteFilesTool } from "../tools/write-files.js";
+import { createGrepSearchTool } from "../tools/grep-search.js";
+import { createExecCommandTool } from "../tools/exec-command.js";
 import { loadConfig } from "./config.js";
 
 let registry = createProviderRegistry({ anthropic, openai, google });
@@ -116,8 +116,7 @@ let chat = new MarkdownAI({
 });
 
 if (opts.showConfig) {
-  log.info(`Config:
-${JSON.stringify(config, null, 2)}`);
+  log.info(`Config:\n${JSON.stringify(config, null, 2)}`);
 }
 
 let res = await tryCatch(chat.run());
